backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-elasticsearch==0.1.0
elasticsearch==8.11.1
sentence-transformers==2.2.2
openai==1.10.0
gunicorn==21.2.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1

